update rust && fix lint
