+++ consensus/src/network.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::monitor;
 use crate::{
     counters,
     logging::LogEvent,
@@ -8,7 +9,6 @@ use crate::{
 };
 use anyhow::{anyhow, ensure};
 use aptos_logger::prelude::*;
-use aptos_metrics_core::monitor;
 use aptos_types::{
     account_address::AccountAddress, epoch_change::EpochChangeProof,
     ledger_info::LedgerInfoWithSignatures, validator_verifier::ValidatorVerifier,
