+++ consensus/src/state_computer.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::monitor;
 use crate::{
     block_storage::tracing::{observe_block, BlockStage},
     commit_notifier::CommitNotifier,
@@ -13,7 +14,6 @@ use anyhow::Result;
 use aptos_crypto::HashValue;
 use aptos_infallible::Mutex;
 use aptos_logger::prelude::*;
-use aptos_metrics_core::monitor;
 use aptos_types::{
     account_address::AccountAddress, contract_event::ContractEvent, epoch_state::EpochState,
     ledger_info::LedgerInfoWithSignatures, transaction::Transaction,
