+++ ecosystem/indexer/src/models/collection.rs
@@ -1,6 +1,6 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
-
+#![allow(clippy::extra_unused_lifetimes)]
 use crate::schema::collections;
 use serde::Serialize;
