+++ ecosystem/indexer/src/models/ownership.rs
@@ -1,6 +1,6 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
-
+#![allow(clippy::extra_unused_lifetimes)]
 use crate::schema::ownerships;
 use serde::Serialize;
