+++ network/src/fuzzing.rs
@@ -37,7 +37,7 @@ pub fn fuzz_network_handshake_protocol_exchange(self_handshake: &HandshakeMsg, d
     fake_socket.set_trailing();
 
     // fuzz the network exchange of HandshakeMsg first
-    let _ = block_on(async move {
+    block_on(async move {
         if let Ok(remote_handshake_msg) = exchange_handshake(self_handshake, &mut fake_socket).await
         {
             // then perform the negotiation
