+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.61.0
+1.62.1
